chore: starting to move some dependencies around a little bit
