+++ plugins/scaffolder-react/package.json
@@ -43,7 +43,21 @@
     "@backstage/version-bridge": "workspace:^",
     "@material-ui/core": "^4.12.2",
     "@material-ui/icons": "^4.9.1",
-    "@material-ui/lab": "4.0.0-alpha.57"
+    "@material-ui/lab": "4.0.0-alpha.57",
+    "@rjsf/core": "^3.2.1",
+    "@rjsf/core-v5": "npm:@rjsf/core@^5.0.0-beta.14",
+    "@rjsf/material-ui": "^3.2.1",
+    "@rjsf/material-ui-v5": "npm:@rjsf/material-ui@^5.0.0-beta.14",
+    "@rjsf/utils": "^5.0.0-beta.14",
+    "@rjsf/validator-ajv6": "^5.0.0-beta.14",
+    "@types/json-schema": "^7.0.9",
+    "classnames": "^2.2.6",
+    "json-schema": "^0.4.0",
+    "json-schema-library": "^7.3.9",
+    "lodash": "^4.17.21",
+    "qs": "^6.9.4",
+    "zod": "~3.18.0",
+    "zod-to-json-schema": "~3.18.0"
   },
   "peerDependencies": {
     "@types/react": "^16.13.1 || ^17.0.0",
