+++ plugins/scaffolder-react/src/next/components/Stepper/Stepper.test.tsx
