+++ plugins/scaffolder-react/src/next/components/Stepper/Stepper.tsx
@@ -29,14 +29,14 @@ import {
 import { type IChangeEvent, withTheme } from '@rjsf/core-v5';
 import { ErrorSchema, FieldValidation } from '@rjsf/utils';
 import React, { useCallback, useMemo, useState } from 'react';
-import { NextFieldExtensionOptions } from '../../../extensions';
+import { NextFieldExtensionOptions } from '../../extensions';
 import { TemplateParameterSchema } from '../../../types';
 import { createAsyncValidators } from './createAsyncValidators';
-import { useTemplateSchema } from './useTemplateSchema';
+import { useTemplateSchema } from '../../hooks/useTemplateSchema';
 import { ReviewState } from './ReviewState';
 import validator from '@rjsf/validator-ajv6';
-import { selectedTemplateRouteRef } from '@backstage/plugin-scaffolder-react';
-import { useFormData } from './useFormData';
+import { selectedTemplateRouteRef } from '../../../routes';
+import { useFormData } from '../../hooks/useFormData';
 import { FormProps } from '../../types';
 
 const useStyles = makeStyles(theme => ({
@@ -66,6 +66,9 @@ export type StepperProps = {
 // of the re-writing we're doing. Once we've migrated, we can import this the exact same as before.
 const Form = withTheme(require('@rjsf/material-ui-v5').Theme);
 
+/**
+ * @alpha
+ */
 export const Stepper = (props: StepperProps) => {
   const { templateName } = useRouteRefParams(selectedTemplateRouteRef);
   const analytics = useAnalytics();
