+++ plugins/scaffolder-react/src/next/components/Stepper/index.ts
@@ -13,8 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-export * from './routes';
-export * from './extensions';
-
-export * from './next';
+export { Stepper } from './Stepper';
