+++ plugins/scaffolder-react/src/next/components/Stepper/schema.ts
