+++ plugins/scaffolder-react/src/next/extensions/types.ts
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 import { ApiHolder } from '@backstage/core-plugin-api';
-import { FieldValidation, FieldProps } from '@rjsf/core';
 import {
   UIOptionsType,
   FieldProps as FieldPropsV5,
@@ -22,62 +21,8 @@ import {
   FieldValidation as FieldValidationV5,
 } from '@rjsf/utils';
 import { PropsWithChildren } from 'react';
-import { JSONSchema7 } from 'json-schema';
 import { JsonObject } from '@backstage/types';
-
-/**
- * Field validation type for Custom Field Extensions.
- *
- * @public
- */
-export type CustomFieldValidator<TFieldReturnValue> = (
-  data: TFieldReturnValue,
-  field: FieldValidation,
-  context: { apiHolder: ApiHolder },
-) => void | Promise<void>;
-
-/**
- * Type for the Custom Field Extension schema.
- *
- * @public
- */
-export type CustomFieldExtensionSchema = {
-  returnValue: JSONSchema7;
-  uiOptions?: JSONSchema7;
-};
-
-/**
- * Type for the Custom Field Extension with the
- * name and components and validation function.
- *
- * @public
- */
-export type FieldExtensionOptions<
-  TFieldReturnValue = unknown,
-  TInputProps = unknown,
-> = {
-  name: string;
-  component: (
-    props: FieldExtensionComponentProps<TFieldReturnValue, TInputProps>,
-  ) => JSX.Element | null;
-  validation?: CustomFieldValidator<TFieldReturnValue>;
-  schema?: CustomFieldExtensionSchema;
-};
-
-/**
- * Type for field extensions and being able to type
- * incoming props easier.
- *
- * @public
- */
-export interface FieldExtensionComponentProps<
-  TFieldReturnValue,
-  TUiOptions extends {} = {},
-> extends FieldProps<TFieldReturnValue> {
-  uiSchema: FieldProps['uiSchema'] & {
-    'ui:options'?: TUiOptions;
-  };
-}
+import { CustomFieldExtensionSchema } from '../../extensions';
 
 /**
  * Type for Field Extension Props for RJSF v5
