+++ plugins/scaffolder-react/src/next/hooks/useTemplateSchema.test.tsx
