+++ plugins/scaffolder-react/src/next/hooks/useTemplateSchema.ts
