+++ plugins/scaffolder-react/src/next/index.ts
@@ -13,8 +13,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-export * from './routes';
+export * from './components';
 export * from './extensions';
-
-export * from './next';
