+++ plugins/scaffolder-react/src/types.ts
@@ -13,8 +13,3 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-export * from './routes';
-export * from './extensions';
-
-export * from './next';
